+++ CHANGELOG.md
@@ -2,10 +2,10 @@
 All notable changes to the "emacs-mcx" extension will be documented in this file.
 
 ## [Unreleased]
+
+## [0.41.2] - 2022-03-26
 ### Fix
 - An error on `isearch-exit`, #1272.
-
-### Fix
 - Restore the find widget string, #1271.
 
 ## [0.41.1] - 2022-03-21